Create a seam for selecting different asset pipeline generators
